requirements.txt
gymnasium-robotics
packaging
stable-baselines3[extra]